RA-84: weight values are now float

+++ app/Filament/Resources/WeightResource/Schemas/WeightForm.php
@@ -25,7 +25,7 @@ class WeightForm extends AbstractFormSchema
                     ->required()
                     ->suffix(Tenant::getWeightUnitLabel())
                     ->minValue(1)
-                    ->integer(),
+                    ->numeric(),
 
                 DatePicker::make('measured_at')
                     ->label('Measured Date')
